fix(socket/rep): correct cleanup on FIN
When a client is done sending data, it sends a FIN message to the
server. This is interpreted by the `Framed` abstraction as an EOF, which
means that when we poll it returns `Poll::Ready(None)`.

The current logic decides to close to connection on such event, but this
is incorrect. The peer is still available to receive a response to the
last data that has been sent. This has been now changed so that we try
to send and flush all messages in the egress queue.

However, pending messages which haven't been handled yet by the consumer
of this library are still dropped. We can refine it later